refactor: category entity 파일 위치 수정
